relocate keycloak service to standalone module
