+++ apps/api/src/modules/keycloak/keycloak.module.ts
@@ -1,0 +1,9 @@
+import { HttpModule } from '@nestjs/axios';
+import { Module } from '@nestjs/common';
+import { KeycloakService } from './keycloak.service';
+
+@Module({
+  imports: [HttpModule],
+  providers: [KeycloakService],
+})
+export class KeycloakModule {}
