+++ apps/api/src/modules/user/user.module.ts
@@ -2,16 +2,17 @@ import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
 import { CrmService } from '../external/crm.service';
 import { ExternalModule } from '../external/external.module';
-import { KeycloakService } from '../external/keycloak.service';
 import { PeoplesoftV2Service } from '../external/peoplesoft-v2.service';
 import { PeoplesoftService } from '../external/peoplesoft.service';
+import { KeycloakModule } from '../keycloak/keycloak.module';
+import { KeycloakService } from '../keycloak/keycloak.service';
 import { PrismaModule } from '../prisma/prisma.module';
 import { PrismaService } from '../prisma/prisma.service';
-import { UserService } from './user.service';
 import { UserResolver } from './user.resolver';
+import { UserService } from './user.service';
 
 @Module({
-  imports: [ExternalModule, HttpModule, PrismaModule],
+  imports: [ExternalModule, HttpModule, KeycloakModule, PrismaModule],
   providers: [
     CrmService,
     KeycloakService,
